+++ src/mainMenu.cpp
@@ -1,13 +1,11 @@
 #include "mainMenu.hpp"
 
-float mainMenuHeight = 640.0;
-float mainMenuWidth = 1024.0;
+unsigned int mainMenuHeight = 640;
+unsigned int mainMenuWidth = 1024;
 
 sf::Texture mainMenuBackground;
 sf::RectangleShape mainMenu;
 
 sf::Texture titleTexture;
-// sf::Sprite title;
 
 sf::Texture playButtonTexture;
-// sf::Sprite playButton;
